backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyMuPDF
Pillow
pytesseract